==> global_vac_index/aim_2/02_construct_index/06_validate_index_markdown.Rmd <==
---
title: "07_validate_index_third_version"
author: "Francisco Rios Casas"
date: "01/10/2023"
output: html_document
---
```{r}
# clear environment
# rm(list=ls())

source(paste0("C:/Users/frc2/Documents/uw-phi-vax/global_vac_index/aim_2/third_version/01_set_up_R.R"))

# library(sklearn)
library(DescTools)

# load index dataset
index_data <- readRDS(paste0(prepped_data_dir, "aim_2/19_index_results_third_version.RDS"))

# load vaccination coverage dataset
vax_data <- readRDS(paste0(prepped_data_dir, "aim_1/01_vaccine_trends.RDS"))

# reshape the vaccination data
vax_data <- pivot_wider(vax_data,
            id_cols=c("location_id", "location_name", "year_id"),
            names_from = "vaccine_name",
            values_from = c(prop_val, prop_upper, prop_lower))

# merge together files for each location and year
full_data <- index_data %>% left_join(vax_data, by=c("gbd_location_id"="location_id", "year"="year_id", "location"="location_name"))

# subset columns
full_data <- full_data %>% select(location, region, year, gbd_location_id, iso_code, iso_num_code, result, prop_val_MCV1, prop_val_DTP1, prop_val_DTP3, prop_val_MCV2) %>% filter(year!=2020)

# recode the region variable into super-regions: 
full_data <- full_data %>% mutate(super_region = case_when(
  region == "Western Europe"            ~ "High Income",
  region == "Southern Latin America"    ~ "High Income",
  region == "High-income North America" ~ "High Income",
  region == "High-income Asia Pacific"  ~ "High Income",
  region == "Australasia"  ~ "High Income",
  
  region == "Tropical Latin America"   ~ "Latin America & Caribbean",
  region == "Caribbean"                ~ "Latin America & Caribbean",
  region == "Andean Latin America"     ~ "Latin America & Caribbean",
  region == "Central Latin America"    ~ "Latin America & Caribbean",
  
  region == "Western Sub-Saharan Africa"  ~ "Sub-Saharan Africa",
  region == "Southern Sub-Saharan Africa" ~ "Sub-Saharan Africa",
  region == "Central Sub-Saharan Africa"  ~ "Sub-Saharan Africa",
  region == "Eastern Sub-Saharan Africa"  ~ "Sub-Saharan Africa",
  
  region == "North Africa and Middle East" ~ "North Africa & Middle East",
  
  region == "Southeast Asia" ~ "South East Asia, East Asia & Oceania",
  region == "Oceania"        ~ "South East Asia, East Asia & Oceania",
  region == "East Asia"      ~ "South East Asia, East Asia & Oceania",
  
  region == "South Asia" ~ "South Asia",
  
  region == "Central Asia"   ~ "Central Europe, Eastern Europe & Central Asia",
  region == "Central Europe" ~ "Central Europe, Eastern Europe & Central Asia",
  region == "Eastern Europe" ~ "Central Europe, Eastern Europe & Central Asia",
  
  TRUE ~ "missing"
))


########################################
##### VALIDATING INDEX #####
########################################

data_subset <- full_data %>% filter(year <2019)

# create a training dataset
row.number <- sample(1:nrow(data_subset), 0.8*nrow(data_subset))
train <- data_subset[row.number,]
test <- data_subset[-row.number,]

# fit a model for vaccines in the complete vaccination series
modeldtp <- glm(prop_val_DTP3~factor(super_region) + year + result, data=train, family = "binomial")
summary(modeldtp)
par(mfrow=c(2,2))
plot(modeldtp)

# predict vaccination levels using the test data
pred1 <- predict(modeldtp, newdata = test, type = "response")

par(mfrow=c(1,1))
plot(pred1, test$prop_val_DTP3,
     xlab='Predicted Vaccination Coverage',
     ylab='Actual Vaccination Coverage',
     main='Predicted vs. Actual Values DTP3, 1990-2018')
abline(a=0, b=1)

# calculate Brier Score: to measure accuracy of probabilistic predictions
BrierScore(test$prop_val_DTP3, pred1)

# fit a second model for measles: MCV2 contained a lot of missing data
modelmcv <- glm(prop_val_MCV1~factor(super_region) + year + result, data=train, family = "binomial")
summary(modelmcv)
par(mfrow=c(2,2))
plot(modelmcv)


# predict the actual vaccination levels using the testing data
pred2 <- predict(modelmcv, newdata = test, type = "response")

par(mfrow=c(1,1))
plot(pred2, test$prop_val_MCV1,
     xlab = "Predicted Vaccination Coverage",
     ylab = "Actual Vaccination Coverage", 
     main = "Predicted vs. Actual Values MCV1, 1990-2018")

# calculate the Brier Score: to measure accuracy of probabilistic predictions
BrierScore(test$prop_val_MCV1, pred2)

########################################
##### Use the model to predict 
##### vaccination coverage in 2019
########################################

newdata <- with(full_data, data.frame(location = rep(unique(full_data$location), each=1, length.out=175),
                                      year = rep(2019, length.out=175)))

newdata <- newdata %>% left_join(full_data, by=c("location", "year"))

pred2019dtp3 <- predict(modeldtp, newdata=newdata, type = "response")
plot(pred2019dtp3, newdata$prop_val_DTP3)
plot(pred2019dtp3, newdata$prop_val_DTP3,
     xlab='Predicted Vaccination Coverage',
     ylab='Actual Vaccination Coverage',
     main='Predicted vs. Actual Values DTP3, 2019')
abline(a=0, b=1)

# calculate the Brier Score
BrierScore(newdata$prop_val_DTP3, pred2019dtp3)

pred2019mcv <- predict(modelmcv, newdata = newdata, type = "response")
plot(pred2019mcv, newdata$prop_val_MCV1,
     xlab='Predicted Vaccination Coverage',
     ylab='Actual Vaccination Coverage',
     main='Predicted vs. Actual Values MCV, 2019')
abline(a=0, b=1)

# calculate the Brier Score
BrierScore(newdata$prop_val_MCV1, pred2019dtp3)


# pred2019dtp1 <- predict(model2, newdata=newdata, type = "response")
# plot(pred2019dtp1, newdata$prop_val_DTP1,
#      xlab='Predicted Vaccination Coverage',
#      ylab='Actual Vaccination Coverage',
#      main='Predicted vs. Actual Values DTP1, 2019')
# abline(a=0, b=1)

# BrierScore(modelmcv)

# fit a model for each vaccine
# model1 <- glm(prop_val_MCV1~factor(region)+year+result, data=train, family = "binomial")
# summary(model1)
# par(mfrow=c(2,2))
# plot(model1)

# evaluate the success of the model 
# pred1 <- predict(model1, newdata = test, type = "response")
# BrierScore(model1)
# par(mfrow=c(1,1))
# plot(pred1, test$prop_val_MCV1,
#      xlab='Predicted Vaccination Coverage',
#      ylab='Actual Vaccination Coverage',
#      main='Predicted vs. Actual Values MCV, 1990-2018')
# abline(a=0, b=1)
# BrierScore(test$prop_val_MCV1, pred1)


# fit second model
# model2 <- glm(prop_val_DTP1~factor(region)+year+result, data=train, family = "binomial")
# summary(model2)
# par(mfrow=c(2,2))
# plot(model2)

# # evaluate the predictions of the model 
# pred2 <- predict(model2, newdata = test, type = "response")
# BrierScore(model2)
# par(mfrow=c(1,1))
# plot(test$prop_val_DTP1, pred2)
# 
# # fit third model
# model3 <- glm(prop_val_DTP3~factor(region)+year+result, data=train, family = "binomial")
# summary(model3)
# par(mfrow=c(2,2))
# plot(model3)

# evaluate the predictions of the model 
# pred3 <- predict(model3, newdata = test, type = "response")
# BrierScore(model3)
# par(mfrow=c(1,1))
# plot(test$prop_val_DTP3, pred3)



# create a dataframe of all locations and vaccines and include the year 2019
# newdata <- with(data, data.frame(ESTIAP = rep(unique(data$ESTIAP), each=224, length.out=12544),
#                                  YEAR = rep(2007:2020, each=16, length.out=12544),
#                                  RACEETHK_R = factor(rep(1:4, each=4, length.out=12544)),
#                                  INCPOV1 = factor(rep(1:4, length.out=12544))))

# newdata <- with(full_data, data.frame(location = rep(unique(full_data$location), each=1, length.out=175),
#                                       year = rep(2019, length.out=175)))
# 
# newdata <- newdata %>% left_join(full_data, by=c("location", "year"))
# 
# 
# pred2019mcv <- predict(model1, newdata = newdata, type = "response")
# plot(pred2019mcv, newdata$prop_val_MCV1,
#      xlab='Predicted Vaccination Coverage',
#      ylab='Actual Vaccination Coverage',
#      main='Predicted vs. Actual Values MCV, 2019')
# abline(a=0, b=1)
# 
# 
# pred2019dtp1 <- predict(model2, newdata=newdata, type = "response")
# plot(pred2019dtp1, newdata$prop_val_DTP1,
#      xlab='Predicted Vaccination Coverage',
#      ylab='Actual Vaccination Coverage',
#      main='Predicted vs. Actual Values DTP1, 2019')
# abline(a=0, b=1)
# 
# 
# pred2019dtp3 <- predict(model3, newdata=newdata, type = "response")
# plot(pred2019dtp3, newdata$prop_val_DTP3)
# plot(pred2019dtp3, newdata$prop_val_DTP3,
#      xlab='Predicted Vaccination Coverage',
#      ylab='Actual Vaccination Coverage',
#      main='Predicted vs. Actual Values DTP3, 2019')
# abline(a=0, b=1)


######################################## #####
##### VALIDATING INDEX For certain years #####
######################################## #####

# data_subset <- full_data %>% filter(year <2019 & year >2015)
# 
# # create a training dataset
# row.number <- sample(1:nrow(data_subset), 0.8*nrow(data_subset))
# train <- data_subset[row.number,]
# test <- data_subset[-row.number,]
# 
# # fit a model for each vaccine
# model1 <- glm(prop_val_MCV1~factor(region)+year+result, data=train, family = "binomial")
# summary(model1)
# par(mfrow=c(2,2))
# plot(model1)
# 
# # evaluate the success of the model 
# pred1 <- predict(model1, newdata = test, type = "response")
# BrierScore(model1)
# par(mfrow=c(1,1))
# plot(pred1, test$prop_val_MCV1,
#      xlab='Predicted Vaccination Coverage',
#      ylab='Actual Vaccination Coverage',
#      main='Predicted vs. Actual Values MCV, 2015-2018')
# abline(a=0, b=1)



```
